# gridex-ops-256-hardening-runtime-patch-2026-07-12-v5/project-files/docs/openapi/customer-portal-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gridex Customer Portal and Website API",
    "version": "2026-07-12",
    "description": "Kanoniskt kontrakt för samtliga /api/v1-routes. API-token bestämmer company_id. company_id accepteras aldrig som fri tenantväljare."
  },
  "servers": [
    {
      "url": "https://app.gridex.se"
    }
  ],
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "getApiV1Events",
        "summary": "Hämta tenantens domänhändelser.",
        "security": [
          {
            "bearerAuth": [
              "events.read"
            ]
          }
        ],
        "x-required-scopes": [
          "events.read"
        ],
        "x-idempotency-required": false,
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Lyckat svar.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "postApiV1Events",
        "summary": "Skicka kundhändelse från hemsida eller portal.",
        "security": [
          {
            "bearerAuth": [
              "website_events.write"
            ]
          }
        ],
        "x-required-scopes": [
          "website_events.write"
        ],
        "x-idempotency-required": true,
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lyckat svar.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Gridex API token"
      }
    },
    "parameters": {
      "RequestId": {
        "name": "x-request-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 200
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 200,
          "pattern": "^[A-Za-z0-9._:+~-]+$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Ogiltig JSON eller validering.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Token saknas eller är ogiltig.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Scope, IP eller origin tillåts inte.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Tvetydig matchning eller idempotenskonflikt.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Affärsvalidering misslyckades.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit överskriden.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internt fel utan databasdetaljer.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "code",
                "message"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "stage": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "field": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "hint": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "request_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}